LPS-91884 Applying same pattern as in the othe preview modules
